Added skeleton for Landing Page tests (screen + test)
